pinctrl: reorganise example system into board family
