+++ examples/pinctrl/building_example/imx/build.sh
