+++ examples/pinctrl/building_example/imx/dts/odroidc4.dts
