+++ examples/pinctrl/building_example/imx/pinctrl.mk
